Move main logic to generator and Refactor
