+++ .gitignore
@@ -1,5 +1,5 @@
 # Binary Files
-firestore-repo
+./bin/firestore-repo
 statik/
 
 # etc
