+++ cmd/firestore-repo/version.go
@@ -1,0 +1,4 @@
+package main
+
+// AppVersion - application version
+const AppVersion = "1.12.0"
