--- constant.go
@@ -1,27 +1,0 @@
-package main
-
-const (
-	equal          = "equal"
-	like           = "like"
-	prefix         = "prefix"
-	suffix         = "suffix"
-	biunigrams     = "Biunigrams"
-	indexLabel     = "IndexLabel"
-	typeString     = "string"
-	typeInt        = "int"
-	typeInt64      = "int64"
-	typeFloat64    = "float64"
-	typeBool       = "bool"
-	typeTime       = "time.Time"
-	typeTimePtr    = "*time.Time"
-	typeLatLng     = "latlng.LatLng"
-	typeReference  = "firestore.DocumentRef"
-	typeMap        = "map[string]"
-	typeStringMap  = "map[string]string"
-	typeIntMap     = "map[string]int"
-	typeInt64Map   = "map[string]int64"
-	typeFloat64Map = "map[string]float64"
-	typeBoolMap    = "map[string]bool"
-	ngType         = "NG"
-	defaultMockOut = "mock/mock_{{ .GeneratedFileName }}/mock_{{ .GeneratedFileName }}.go"
-)
