+++ generator/templates/constant.go.tmpl
