+++ generator/templates/errors.go.tmpl
