+++ generator/templates/label.go.tmpl
