+++ generator/templates/misc.go.tmpl
