+++ generator/templates/unique.go.tmpl
