+++ generator/testfiles/auto/lock.go
