+++ generator/testfiles/auto/lock_meta.go
