+++ generator/testfiles/auto/task.go
