+++ generator/testfiles/auto/tests/model_test.go
