+++ generator/testfiles/not_auto/tests/model_test.go
